Flickering and Routing fixes removed e2etest node env. simply uses test now fixed issue with myblogs page not loading on refresh all non api routes now go to static frontend index.js Fixed flickering due to use of useBreakpoint Pulled MobileMenu.js and NavBarLinks.js out of NavBar.js
