+++ backend/app.js
@@ -2,6 +2,7 @@ const express = require("express");
 require("express-async-errors");
 const cors = require("cors");
 const mongoose = require("mongoose");
+const path = require("path");
 const middleware = require("./utils/middleware");
 const logger = require("./utils/logger");
 const blogsRouter = require("./controllers/blogs");
@@ -35,11 +36,15 @@ app.use("/api/blogs", middleware.userExtractor, blogsRouter);
 app.use("/api/users", usersRouter);
 app.use("/api/login", loginRouter);
 
-if (process.env.NODE_ENV === "e2etest") {
+if (process.env.NODE_ENV !== "production") {
   // eslint-disable-next-line global-require
   const testingRouter = require("./controllers/testing");
   app.use("/api/testing", testingRouter);
 }
+
+app.get("/*", (req, res) => {
+  res.sendFile(path.join(__dirname, "build", "index.html"));
+});
 
 app.use(middleware.unknownEndpoint);
 app.use(middleware.errorHandler);
